requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
Django==5.2.4
whitenoise==6.6.0
gunicorn==21.2.0
django-environ==0.12.0
django-multiselectfield==1.0.1
django-heroku
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.98.0
outcome==1.3.0.post0
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PySocks==1.7.1
python-dotenv==1.1.1
reportlab==4.4.3
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
selenium
bs4==0.0.1
mysqlclient==2.2.4
PyMySQL==1.1.0
boto3
django-storages
requests
